bots/SpotifyBot/requirements.txt
anyio==4.0.0
APScheduler==3.6.3
async-timeout==4.0.3
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
python-telegram-bot==13.7
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
tornado==6.3.3
tzlocal==5.2
urllib3==2.1.0
